sexp: remove package
This package was broken by the removal of undocumented API headers in
the R distribution. Sadly this greatly diminishes the value of rgo since
the performance of the documented approach requiring interchange via Cgo
is significantly worse than the direct memory access that this package
provided. It also great reduces the flexibility of operations that could
be performed using rgo.

The change that breaks this is git-svn-id: https://svn.r-project.org/R/trunk@80644
00db46b3-68df-0310-9c12-caf00c1e9a41.